--- a/contrib/tvmop/compile.py
+++ b/contrib/tvmop/compile.py
@@ -18,12 +18,15 @@
 # coding: utf-8
 """TVM Operator compile entry point"""
 import tvm
+from tvm import autotvm
 
 import os
 import argparse
 import re
+import json
 import logging
 from tvmop.opdef import __OP_DEF__
+from tvmop.space import ConfigSpaces, ConfigSpace
 from tvm.autotvm.measure.measure_methods import set_cuda_target_arch
 
 logging.basicConfig(level=logging.INFO)
@@ -70,20 +73,52 @@ if __name__ == "__main__":
                         help="Target path which stores compiled library")
     parser.add_argument('--cuda-arch', type=str, default=None, dest='cuda_arch',
                         help='The cuda arch for compiling kernels for')
+    parser.add_argument("--config", action="store", required=True, dest="config_path",
+                        help="Path which stores the config file")
     arguments = parser.parse_args()
 
     func_list_llvm = []
     func_list_cuda = []
+    config_spaces = ConfigSpaces()
 
     # TODO: attach instruction features to the library, e.g., avx-512, etc.
-    for operator_def in __OP_DEF__:
-        for sch, args, name in operator_def.invoke_all():
-            if tvm.module.enabled(get_target(operator_def.target)):
-                func_list = func_list_llvm if operator_def.target == "cpu" else func_list_cuda
-                func_lower = tvm.lower(sch, args,
-                                       name=name,
-                                       binds=operator_def.get_binds(args))
-                func_list.append(func_lower)
+    for op in __OP_DEF__:
+        if tvm.module.enabled(get_target(op.target)):
+            func_list = func_list_llvm if op.target == "cpu" else func_list_cuda
+            for each_kwargs in op.arg_combination:
+                if (op.attrs_valid(**each_kwargs)):
+                    name = op.name \
+                        + ''.join(["{}_{}".format(key, each_kwargs[key]) for key in op.attrs])
+                    if op.dispatch is True:
+                        config_space = autotvm.ConfigSpace()
+                        with autotvm.task.ApplyConfig(config_space):
+                            sch, args = op.func(fallback=False, **each_kwargs)
+                        # register dispatch schedules
+                        for i in range(len(config_space)):
+                            config_entity = config_space.get(i)
+                            with autotvm.task.ApplyConfig(config_entity):
+                                sch, args = op.func(fallback=False, **each_kwargs)
+                            subname = name + "index_" + str(i) + \
+                                ''.join(["%s_%d" % (arg.dtype, len(arg.shape)) for arg in args])
+                            func_lower = tvm.lower(sch, args,
+                                                name=subname,
+                                                binds=op.get_binds(args))
+                            func_list.append(func_lower)
+                        # register config space
+                        config_spaces[name] = ConfigSpace.from_tvm(config_space)
+                        # register fallback schedule
+                        config_space = autotvm.ConfigSpace()
+                        with autotvm.task.ApplyConfig(config_space):
+                            sch, args = op.func(fallback=True, **each_kwargs)
+                        subname = name + "fallback" + \
+                            ''.join(["%s_%d" % (arg.dtype, len(arg.shape)) for arg in args])
+                        func_lower = tvm.lower(sch, args, name=subname, binds=op.get_binds(args))
+                        func_list.append(func_lower)
+                    else:
+                        sch, args = op.func(**each_kwargs)
+                        subname = name + ''.join(["%s_%d" % (arg.dtype, len(arg.shape)) for arg in args])
+                        func_lower = tvm.lower(sch, args, name=subname, binds=op.get_binds(args))
+                        func_list.append(func_lower)
 
     lowered_funcs = {get_target("cpu"): func_list_llvm}
     if len(func_list_cuda) > 0:
@@ -96,3 +131,5 @@ if __name__ == "__main__":
             set_cuda_target_arch(cuda_arch)
     func_binary = tvm.build(lowered_funcs, name="tvmop")
     func_binary.export_library(arguments.target_path)
+    with open(arguments.config_path, "w") as f:
+        json.dump(config_spaces.to_json_dict(), f)