--- a/contrib/tvmop/opdef.py
+++ b/contrib/tvmop/opdef.py
@@ -77,12 +77,41 @@ class OpDef:
     def invoke_all(self):
         for each_kwargs in self.arg_combination:
             if self.attrs_valid(**each_kwargs):
-                sch, args = self.func(**each_kwargs)
                 name = self.name \
-                    + ''.join(["{}_{}".format(key, each_kwargs[key]) for key in self.attrs]) \
-                    + ''.join(["%s_%d" % (arg.dtype, len(arg.shape))
-                               for arg in args if hasattr(arg, 'shape')])
-                yield sch, args, name
+                    + ''.join(["{}_{}".format(key, each_kwargs[key]) for key in self.attrs])
+                if self.dispatch is False:
+                    sch, args = self.func(**each_kwargs)
+                    yield sch, args, name
+                else:
+                    # register dispatch schedules
+                    config_space = autotvm.ConfigSpace()
+                    with autotvm.task.ApplyConfig(config_space):
+                            sch, args = self.func(fallback=False, **each_kwargs)
+                    for i in range(len(config_space)):
+                        config_entity = config_space.get(i)
+                        with autotvm.task.ApplyConfig(config_entity):
+                            sch, args = self.func(fallback=False, **each_kwargs)
+                        subname = name + "index_" + str(i)
+                        yield sch, args, subname
+                    # register fallback schedule
+                    config_space = autotvm.ConfigSpace()
+                    with autotvm.task.ApplyConfig(config_space):
+                            sch, args = self.func(fallback=True, **each_kwargs)
+                    subname = name + "fallback"
+                    yield sch, args, subname
+
+    def get_op_name(self, name, args):
+        return name + ''.join(["%s_%d" % (arg.dtype, len(arg.shape)) for arg in args if hasattr(arg, 'shape')])
+
+    def get_config_spaces(self):
+        for each_kwargs in self.arg_combination:
+            if self.attrs_valid(**each_kwargs) and self.dispatch is True:
+                name = self.name \
+                    + ''.join(["{}_{}".format(key, each_kwargs[key]) for key in self.attrs])
+                config_space = autotvm.ConfigSpace()
+                with autotvm.task.ApplyConfig(config_space):
+                    self.func(fallback=False, **each_kwargs)
+                yield config_space, name
 
     def get_binds(self, args):
         if self.auto_broadcast: