--- a/python/mxnet/__init__.py
+++ b/python/mxnet/__init__.py
@@ -87,6 +87,8 @@ from . import rnn
 
 from . import gluon
 
+from . import tvmop
+
 __version__ = base.__version__
 
 # Dist kvstore module which launches a separate process when role is set to "server".