--- a/python/mxnet/space.py
+++ b/python/mxnet/space.py
@@ -1,0 +1,1 @@
+../../contrib/tvmop/space.py