@hzfan 2019-09-06
> Just for curious. Based on my knowledge, tvm op kernel is pre-compiled and then linked together with MXNet. How can it be configured according to the runtime input shapes?

Yes, kernels are pre-compiled. At compile time, several different schedules (kernels) for a single op are defined and compiled. Then at runtime, with the runtime input shape, the most suitable kernel is chosen.

It's true that the kernel is pre-compiled, but we have multiple available kernels for one single op, so we can choose the most efficient one based on the runtime input shape.