@ZhennanQin 2019-09-09
@hzfan Thanks for explanation.  My next question is, how do I know which schedule is the best one for a certain input shape? Static rule defined or runtime tuned?