@hzfan 2019-09-09
> @hzfan Thanks for explanation. My next question is, how do I know which schedule is the best one for a certain input shape? Static rule defined or runtime tuned?

That's a good question. Actually we have considered both options, and for now we use simple static rules. To be more specific, for now I require the size of a for-loop to be multiples of its splitting factor (if the for-loop is splitted). This helps eliminate a if-condition, and thus makes it faster.

Runtime tuning has also been considered, but has not been implemented in this version. The idea is to try all the available schedules for every runtime shape, measure their performance, and cache the best choice. This is quite similar to autotvm.