@hzfan 2019-12-23
> do we have a developer guide using tvm op?

Seems @yzhliu is working on it.